# Improved parallelism in osregex execute and rule engine