|Related issue|
|---|
|Closes #7938 |

Hi Team!,

This PR is small but brings big performance improvements to Analaysisd. It ensures parallelism at the time of decoding and rule matching, improving the performance of version 4.1 by up to 262% with 16 threads.

Rule trees and decoders trees that include regexes are shared between threads. The execution of regexes involves taking a semaphore that is part of the regex structure, this implies that if a thread takes the semaphore and performs a context switch, the other threads that need to process the same regex will be queued until the first thread returns it.
The very nature of the decoders makes this more likely at the decoding stage than at the matching stage.

The solution is only to lock the semaphore in case you are not using the parallelism mechanism of the OSRegex_Execute_ex function and force the rules engine to always use this mechanism.

For the tests we used the `gettimeofday()` function that allows us to measure the time it takes to perform actions by co-considering the CPU-off. The following times were measured:

- **Decode time**: It is the time it takes to decode the event. (Specifically the time taken by the `DecodeEvent` function)
- **Rule match time**: Time it takes to match the rule (The time from the time it tries to match with the first rule until before the event is copied to the queue to be written in the alert.log).
- **OS_Regex time**: It is the time (cumulative) used in the OSRegex_Execute_ex function, both in the decoding phase and in the rule matching phase.
- Number of times that OSRegex_Execute_ex was called with that event (If we divide it by `OS_Regex time` gives us the average time it takes to make a match for that event),

The tests were performed with the same number of threads for decoding and rule matching. 1, 2, 4 and 8 threads each, on a 16-thread virtual machine.

In a test on the master branch (v4.2) 8910 logs were sent (10 times the same set of 981 logs) and the following results were obtained

![master](https://user-images.githubusercontent.com/1891958/112494539-e64bbf00-8d61-11eb-9239-8cef7740b4c7.png)

The time it takes is from 56us (average time between decode and match) with one thread, up to 252us for 16 threads.
With the optimization of this PR, better results were obtained, from 54us with one thread to 94us with 16 threads.

![PR](https://user-images.githubusercontent.com/1891958/112494514-e1870b00-8d61-11eb-841c-1fce13983a4f.png)

A time comparison shows their improvement

![PR vs Master](https://user-images.githubusercontent.com/1891958/112494491-daf89380-8d61-11eb-9e3c-f7157d05545a.png)


Attached is what is needed to replicate the tests.
[analysisd_performance.zip](https://github.com/wazuh/wazuh/files/6185194/analysisd_performance.zip) contains a set of 891 logs for testing, a go script to send the data from different locations and a bash script (very badly programmed) that runs Wazuh, change its configuration and reports the times by extracting the data in alerts.log.


Instructions:
- Apply the patches (`0002-First-optimeze.patch` may not be applied, it is the patch with optimizations).
- Adjust the paths in `test_per.sh`.
- Run `test_per.sh` and wait a few minutes, it will generate all the necessary results for that branch.



## Tests

  - [x] IT
  - [x] UT
  
  
<!-- Minimum checks required -->
- Compilation without warnings in every supported platform
  - [x] Linux
  - [ ] Windows
  - [ ] MAC OS X
- [x] Source installation
- [x] Package installation
- [x] Source upgrade
- [ ] Package upgrade
- [ ] Review logs syntax and correct language
- [ ] QA templates contemplate the added capabilities

<!-- Depending on the affected OS -->
- Memory tests for Linux
  - [x] Scan-build report
  - [x] Coverity
  - [x] Valgrind in logtest (memcheck and descriptor leaks check)
  - [x] Valgrind in Analysisd (memcheck and descriptor leaks check)
  - [x] Valgrind in Analysisd and logtest at the same time (memcheck and descriptor leaks check)
  - [x] Helgrind (`parallel-regex` tool)
  - [ ] Dr. Memory
  - [ ] AddressSanitizer
- Memory tests for Windows
  - [ ] Scan-build report
  - [ ] Coverity
  - [ ] Dr. Memory
- Memory tests for macOS
  - [ ] Scan-build report
  - [ ] Leaks
  - [ ] AddressSanitizer

<!-- Checks for huge PRs that affect the product more generally -->
- [x] Retrocompatibility with older Wazuh versions
- [ ] Working on cluster environments
- [ ] Configuration on demand reports new parameters
- [x] The data flow works as expected (agent-manager-api-app)
- [ ] Added unit tests (for new features)
- [x] Stress test for affected components